Remove privacy policy step from callback sign up
Remove the privacy policy step from callback sign up; the policy is set on the
first step instead.
