+++ config/locales/en.yml
@@ -304,11 +304,6 @@ en:
               invalid: Enter a valid phone number
             phone_call_scheduled_at:
               blank: "You need to choose a time to call"
-        callbacks/steps/privacy_policy:
-          attributes:
-            accept_privacy_policy:
-              blank: Accept the privacy policy to continue
-              accepted: Accept the privacy policy to continue
         callbacks/steps/talking_points:
           attributes:
             talking_points:
@@ -476,10 +471,6 @@ en:
       events_steps_further_details:
         subscribe_to_mailing_list: |-
           Would you like to receive email updates to help you get into teaching? If you say no, you will still receive confirmation emails about any events you sign up to attend.
-      callbacks_steps_privacy_policy:
-        accept_privacy_policy:
-          text: Are you over 16 and do you agree to our %{link}?
-          link: privacy policy
 
   internal:
     status:
