+++ spec/factories/callbacks/personal_details_factory.rb
@@ -3,5 +3,6 @@ FactoryBot.define do
     first_name { "Test" }
     sequence(:last_name) { |n| "User #{n}" }
     sequence(:email) { |n| "testuser#{n}@testing.education.gov.uk" }
+    accepted_policy_id { "abc123" }
   end
 end
