--- spec/factories/callbacks/privacy_policy_factory.rb
@@ -1,5 +1,0 @@
-FactoryBot.define do
-  factory :callbacks_privacy_policy, class: "Callbacks::Steps::PrivacyPolicy" do
-    accept_privacy_policy { "1" }
-  end
-end
