+++ spec/factories/callbacks/talking_points_factory.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :callbacks_talking_points, class: "Callbacks::Steps::TalkingPoints" do
+    talking_points { "Funding your training" }
+  end
+end
