+++ spec/models/callbacks/steps/personal_details_spec.rb
@@ -9,6 +9,7 @@ describe Callbacks::Steps::PersonalDetails do
   it { is_expected.to respond_to :first_name }
   it { is_expected.to respond_to :last_name }
   it { is_expected.to respond_to :email }
+  it { is_expected.to respond_to :accepted_policy_id }
 
   describe "validations" do
     subject { instance.errors.messages }
@@ -18,6 +19,11 @@ describe Callbacks::Steps::PersonalDetails do
     it { is_expected.to include(:first_name) }
     it { is_expected.to include(:last_name) }
     it { is_expected.to include(:email) }
+    it { is_expected.to include(:accepted_policy_id) }
+  end
+
+  describe "accepted_policy_id" do
+    it { is_expected.to validate_presence_of(:accepted_policy_id) }
   end
 
   describe "#first_name" do
