+++ spec/requests/rate_limiting_spec.rb
@@ -2,6 +2,7 @@ require "rails_helper"
 
 describe "Rate limiting", type: :request do
   include_context "with stubbed types api"
+  include_context "with stubbed latest privacy policy api"
   include_context "with stubbed candidate create access token api"
 
   let(:ip) { "1.2.3.4" }
